Add position to compare with faster solutions.
Thanks reddit, the solution to part 1 must be the median, the
solution to part 2 is either floor(mean) or ceil(mean).

In order for part 1 to have a unique solution the median position must
be included in the set of crab submarine positions.

TDOD: faster solutions still to be implemented.